+++ backend/tca-tool/src/main/java/com/backend/tcatool/dto/transformerdata/TransformerDataPutDto.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class TransformerDataPutDto {
-    private Integer transformerId;
+    private long transformerId;
     private String city;
     private String country;
     private InstallationType installationType;
